+++ src/pages/incidents/incidents.ts
@@ -30,11 +30,11 @@ export default class IncidentsPage extends Vue {
   public streamsData: Stream[] | undefined
   public incidentsStatus: IncidentStatus[] = [
     { value: 'any', label: 'Any', checked: true },
-    { value: 'open', label: 'Open', checked: false },
-    { value: 'closed', label: 'Closed', checked: false },
-    // { value: 'recent', label: 'Recent', checked: false },
+    { value: 'recent', label: 'Recent', checked: false },
     { value: 'hot', label: 'Hot', checked: false }
   ]
+
+  public incidentsClosed: IncidentStatus = { value: 'closed', label: 'Include closed incidents', checked: false }
 
   public limit = 2
   public statusSelected = false
@@ -147,6 +147,7 @@ export default class IncidentsPage extends Vue {
     return await StreamService.getStreams({
       projects: [projectId],
       ...status !== undefined && this.optionsForStatus(status),
+      include_closed_incidents: this.incidentsClosed.checked ? true : undefined,
       limit: this.paginationSettings.limit,
       offset: this.paginationSettings.offset * this.paginationSettings.limit,
       keyword: this.searchLabel,
@@ -167,9 +168,7 @@ export default class IncidentsPage extends Vue {
   }
 
   public optionsForStatus (status: string): statusOptions | undefined {
-    if (status === 'closed') return { include_closed_incidents: true }
-    else if (status === 'open') return { include_closed_incidents: false }
-    else if (status === 'hot') return { has_hot_incident: true }
+    if (status === 'hot') return { has_hot_incident: true }
     else if (status === 'recent') return { has_new_events: true }
   }